requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
chromadb==0.4.18
google-generativeai==0.3.2
redis==5.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
httpx==0.25.2
torch>=2.0.0
transformers>=4.30.0
IndicTransToolkit>=1.0.0
